# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  uploader.CompanyFileMetaItem:
    properties:
      created_at:
        type: string
      file_key:
        type: string
      file_name:
        type: string
      file_size:
        type: integer
      file_txn_meta:
        type: string
      file_txn_type:
        type: integer
      id:
        type: string
    type: object
  uploader.CreateUploaderConfigRequest:
    properties:
      aws_access_key:
        type: string
      aws_bucket_name:
        type: string
      aws_bucket_region:
        type: string
      aws_secret_key:
        type: string
      default_quota:
        type: integer
      is_active:
        type: integer
      total_quota:
        type: integer
    type: object
  uploader.CreateUploaderConfigResponse:
    properties:
      created_at:
        type: string
    type: object
  uploader.DeleteFileRequest:
    properties:
      file_key:
        type: string
      file_txn_meta:
        type: string
    type: object
  uploader.DeleteFileResponse:
    properties:
      file_key:
        type: string
    type: object
  uploader.DeleteFolderRequest:
    properties:
      file_txn_meta:
        type: string
      folder_prefix:
        type: string
    type: object
  uploader.DeleteFolderResponse:
    properties:
      deleted_bytes:
        type: integer
      deleted_count:
        type: integer
      folder_prefix:
        type: string
    type: object
  uploader.GenerateUploadURLRequest:
    properties:
      file_name:
        description: required
        type: string
      file_size:
        description: required
        type: integer
      file_txn_meta:
        description: optional
        type: string
      file_txn_type:
        description: required (e.g. 1=upload)
        type: integer
      loc_tag:
        type: string
    type: object
  uploader.GenerateUploadURLResponse:
    properties:
      file_id:
        type: string
      file_key:
        type: string
      upload_url:
        type: string
    type: object
  uploader.ListCompanyFilesResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/uploader.CompanyFileMetaItem'
        type: array
      remaining_quota:
        type: integer
      total_usage_quota:
        type: integer
      used_quota:
        type: integer
    type: object
  uploader.ListFoldersResponse:
    properties:
      items:
        items:
          type: string
        type: array
      next_token:
        type: string
    type: object
host: localhost:9393
info:
  contact: {}
  description: API for Jupyter Platform (DMS, contact-us, etc).
  title: Jupyter API
  version: 1.0.0
paths:
...
  /uploader/config:
    post:
      consumes:
      - application/json
      description: Creates a new uploader_config entry
      parameters:
      - description: Uploader config payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/uploader.CreateUploaderConfigRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/uploader.CreateUploaderConfigResponse'
        "400":
          description: invalid request
          schema:
            type: string
        "500":
          description: internal error
          schema:
            type: string
      summary: Create uploader configuration
      tags:
      - uploader
  /uploader/files:
    get:
      description: Uses X-API-Key to identify company and returns its files_meta records
      parameters:
      - description: Company API key
        in: header
        name: X-API-Key
        required: true
        type: string
      - description: Max number of items (default 50)
        in: query
        name: limit
        type: integer
      - description: Offset for pagination (default 0)
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/uploader.ListCompanyFilesResponse'
        "401":
          description: unauthorized
          schema:
            type: string
        "500":
          description: internal error
          schema:
            type: string
      summary: List files for the calling company
      tags:
      - uploader
    post:
      consumes:
      - application/json
      description: Validates API key, generates a presigned S3 upload URL using company
        AWS config, and stores files_meta
      parameters:
      - description: Company API key
        in: header
        name: X-API-Key
        required: true
        type: string
      - description: File upload request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/uploader.GenerateUploadURLRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/uploader.GenerateUploadURLResponse'
        "400":
          description: invalid request
          schema:
            type: string
        "401":
          description: unauthorized
          schema:
            type: string
        "500":
          description: internal error
          schema:
            type: string
      summary: Generate S3 presigned upload URL and create file meta
      tags:
      - uploader
  /uploader/files/delete:
    post:
      consumes:
      - application/json
      description: Deletes a file from S3 for the authenticated company and records
        a files_meta entry with file_txn_type=2
      parameters:
      - description: Company API key
        in: header
        name: X-API-Key
        required: true
        type: string
      - description: Delete file request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/uploader.DeleteFileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/uploader.DeleteFileResponse'
        "400":
          description: invalid request
          schema:
            type: string
        "401":
          description: unauthorized
          schema:
            type: string
        "500":
          description: internal error
          schema:
            type: string
      summary: Delete a single file by key
      tags:
      - uploader
  /uploader/folders:
    get:
      description: Uses X-API-Key to identify company and returns a paginated list
        of folder prefixes under the company slug
      parameters:
      - description: Company API key
        in: header
        name: X-API-Key
        required: true
        type: string
      - description: Folder prefix (relative to company root) e.g. 'invoices/2023'
        in: query
        name: prefix
        type: string
      - description: Max number of items (default 100)
        in: query
        name: limit
        type: integer
      - description: Continuation token for pagination
        in: query
        name: next_token
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/uploader.ListFoldersResponse'
        "401":
          description: unauthorized
          schema:
            type: string
        "500":
          description: internal error
          schema:
            type: string
      summary: List folders (prefixes) in S3 for the calling company
      tags:
      - uploader
  /uploader/folders/delete:
    post:
      consumes:
      - application/json
      description: Deletes all objects under folder_prefix for the authenticated company
        and records a files_meta entry with file_txn_type=3
      parameters:
      - description: Company API key
        in: header
        name: X-API-Key
        required: true
        type: string
      - description: Delete folder request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/uploader.DeleteFolderRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/uploader.DeleteFolderResponse'
        "400":
          description: invalid request
          schema:
            type: string
        "401":
          description: unauthorized
          schema:
            type: string
        "500":
          description: internal error
          schema:
            type: string
      summary: Delete all files under a folder (prefix)
      tags:
      - uploader
  /uploader/folders/files:
    get:
      description: Uses X-API-Key to identify company and returns files under a specific
        folder prefix
      parameters:
      - description: Company API key
        in: header
        name: X-API-Key
        required: true
        type: string
      - description: Folder prefix (relative to company root)
        in: query
        name: folder
        required: true
        type: string
      - description: Max number of items (default 50)
        in: query
        name: limit
        type: integer
      - description: Offset for pagination (default 0)
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/uploader.ListCompanyFilesResponse'
        "400":
          description: invalid request
          schema:
            type: string
        "401":
          description: unauthorized
          schema:
            type: string
        "500":
          description: internal error
          schema:
            type: string
      summary: List files in a specific folder
      tags:
      - uploader
swagger: "2.0"